+++ src/commands/skip.ts
@@ -1,7 +1,8 @@
 import Command from '../interfaces/command';
-import { Message } from 'discord.js';
+import { Message, Role } from 'discord.js';
 import bot from '..';
 import Messages from '../interfaces/messages';
+import { Player } from 'discord.js-lavalink';
 
 export default class SkipCommand implements Command {
     info = {
@@ -15,36 +16,31 @@ export default class SkipCommand implements Command {
         if(!message.member.voiceChannel)
             message.reply(messages.connectVoice);
 
-        let player = bot.player.manager.get(message.guild.id);
-        let skipping = bot.player.settings[message.guild.id].skipping;
+        let player: Player = bot.player.manager.get(message.guild.id);
+        let skipping: string[] = bot.player.settings[message.guild.id].skipping;
 
-        if (!player) {
-            return message.reply(messages.notPlaying)
-        }
+        if(!player)
+            return message.reply(messages.notPlaying);
 
         if(message.member.voiceChannel.id == player.channel) {
-            if (message.member.roles.find('name', 'DJ') || message.member.permissions.hasPermission("ADMINISTRATOR")) {
-                if (player) {
+            if(message.member.roles.find((r: Role): boolean => r.name == 'DJ')
+               || message.member.permissions.hasPermission('ADMINISTRATOR')) {
+                if(player) {
                     player.stop();
                     return message.reply(messages.skipped);
                 }
             }
 
-            if (skipping.indexOf(message.author.id) == -1) {
+            if(!skipping.includes(message.author.id)) {
                 skipping.push(message.author.id);
-                if(skipping.length >= (message.member.voiceChannel.members.size-1)/2) {
-                    if (player) {
+                if(skipping.length >= (message.member.voiceChannel.members.size - 1) / 2) {
+                    if(player) {
                         player.stop();
                         return message.reply(messages.skipped);
                     }
-                } else {
+                } else
                     message.reply(messages.skippedNeed + ` ${skipping.length} / ${Math.ceil((message.member.voiceChannel.members.size - 1) / 2)}`)
-                }
-            } else {
-                message.reply(messages.skippedAlready)
-            }
-        } else {
-            message.reply(messages.connectVoice);
-        }
+            } else message.reply(messages.skippedAlready)
+        } else message.reply(messages.connectVoice);
     }
 }