+++ src/dashboard/public/components/Commands.vue
@@ -1,57 +1,36 @@
 <template id="commands">
     <div>
-            <div class="command-ct">
-                <h1>Commands</h1>
-                <div class="table">
-                    <div class="item-group" v-for="(cat, i) in categories" :key=i>
-                        <div class="item cat-title">{{ cat.name }}</div>
-                        <div class="item" v-for="(command, i) in cat.commands" :key=i>
-                            <div class="name">
-                                {{ command.names[0] }}
-                            </div>
-                            <div class="desc">
-                                {{ command.description }}
-                            </div>
-                    </div>
+        <div class="command-ct">
+            <h1>Commands</h1>
+            <div class="table">
+                <div class="item-group" v-for="(cat, i) in categories" :key=i>
+                    <div class="item cat-title">{{ cat.name }}</div>
+                    <div class="item appear" v-for="(command, i) in cat.commands" :key=i>
+                        <div class="name">
+                            {{ command.names[0] }}
+                        </div>
+                        <div class="desc">
+                            {{ command.description }}
+                        </div>
                     </div>
                 </div>
             </div>
+        </div>
     </div>
 </template>
 
 <script>
 export default {
-    data: function() {
-        return {
-            categories: [
-                {
-                    name: 'admin',
-                    commands: []
-                },
-                {
-                    name: 'developer',
-                    commands: []
-                },
-                {
-                    name: 'music',
-                    commands: []
-                },
-                {
-                    name: 'music (DJ)',
-                    commands: []
-                },
-                {
-                    name: 'stats',
-                    commands: []
-                },
-                {
-                    name: 'tool',
-                    commands: []
-                }
-            ]
-        }
-    },
-    props: ['dashboard'],
+    data: () => ({
+        categories: [
+            { name: 'admin', commands: [] },
+            { name: 'developer', commands: [] },
+            { name: 'music', commands: [] },
+            { name: 'music (DJ)', commands: [] },
+            { name: 'stats', commands: [] },
+            { name: 'tool', commands: [] }
+        ]
+    }),
     mounted: function() {
         fetch('/api/commands')
             .then(res => res.json())
@@ -60,17 +39,27 @@ export default {
                     if(a.names[0] < b.names[0]) return -1;
                     if(a.names[0] > b.names[0]) return 1;
                     return 0;
-                })
+                });
                 resp.forEach(cmd => {
                     let bb = this.categories.find(cat => cat.name == cmd.category);
                     bb.commands.push(cmd);
-                })
+                });
             });
     }
-}
+};
 </script>
 
 <style scoped>
+@-webkit-keyframes fadeIn {
+    from { opacity: 0; }
+      to { opacity: 1; }
+}
+
+@keyframes fadeIn {
+    from { opacity: 0; }
+      to { opacity: 1; }
+}
+
 h1 {
     font-weight: 300;
     font-size: 3rem;
@@ -86,7 +75,7 @@ h1 {
 .item {
     display: flex;
     flex-direction: row;
-    border-bottom: 1px solid hsla(0,0%,100%,.04);
+    border-bottom: 1px solid hsla(0, 0%, 100%, .04);
 }
 
 .name {