+++ src/dashboard/public/components/Dashboard.vue
@@ -24,15 +24,13 @@ import Radios from './Radios.vue';
 
 export default {
     components: { Menu, Guild, Commands, Radios },
-    data: function() {
-        return {
-            guilds: [],
-            selected: {},
-            commandsSelected: null,
-            view: 0,
-            userData: null, // logged user data
-        }
-    },
+    data: () => ({
+        guilds: [],
+        selected: {},
+        commandsSelected: null,
+        view: 0,
+        userData: null, // logged user data
+    }),
     beforeMount: function() {
         fetch('/api/guilds')
             .then(resp => resp.json())
@@ -47,13 +45,11 @@ export default {
         if(resp.ok) {
             const data = await resp.json();
             this.userData = data.data;
-        } else {
-            this.$router.push('/');
-        }
+        } else this.$router.push('/');
     },
     methods: {
         select(id) {
-            fetch('/api/guild?guild=' + id)
+            fetch(`/api/guild?guild=${id}`)
                 .then(resp => resp.json())
                 .then(data => {
                     this.selected = data;
@@ -67,20 +63,10 @@ export default {
             window.location = '/api/logout';
         }
     }
-}
+};
 </script>
 
 <style lang="scss" scoped>
-@media screen and (min-width: 512px) {
-    .top {
-        flex-direction: row;
-        position: sticky;
-    }
-}
-
-@media screen and (max-width: 512px) {
-    .top { flex-direction: column; }
-}
 .top {
     background: #202225;
     box-shadow: 2px 3px 3px rgba(0,0,0,.24);