+++ CHANGELOG.md
@@ -1,11 +1,23 @@
 # Changelog
 
 # 2.6
+## 2.6.2
+- Added auto update check
+- Updated `weather` command
+
+## 2.6.1
+- Updated dashboard
+- Added goodbyes (`greeting goodbye ...`)
+- Bugfixes
+- Prettied some code
+- Added developer mode and custom dashboard url
+
 ## 2.6.0
 - *Huge* music update
     - Added radios
     - Updated almost everything
     - Replaced ytdl with Lavalink
+    - Added `bass` command
 - Prettied some code
 - Updated `eval, weather` commands